isolate non working vel controlelr
